update delete Member

+++ backend/src/main/java/org/example/backend/service/MemberService.java
@@ -1,6 +1,7 @@
 package org.example.backend.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.backend.exception.ErrorMessage;
 import org.example.backend.exception.IdNotFoundException;
 import org.example.backend.model.Book;
 import org.example.backend.model.Member;
@@ -47,6 +48,10 @@ public class MemberService {
     public void deleteById(String id) {
         if (!memberRepository.existsById(id)) {
             throw new IdNotFoundException(id, "Member");
+        }
+        List<Book> books = getBooksByMemberId(id);
+        if(!books.isEmpty()) {
+            throw new RuntimeException("Member with id " + id + " cannot be deleted");
         }
         memberRepository.deleteById(id);
     }
